Properly escape template IDs when trying to autoload JS classes.
If eventum was installed in a directory with more than a single '-' character in it such as "eventum-test-test" or a template had multiple '-' in the filename and the full template path was used, a JS error would be thrown loading the page. This was not a problem with any of the core eventum code but could be triggered by custom CRM or workflow code.