Change quick filter bar back to plain select boxes
Use plain select boxes in the quick filter bar to avoid taking too much
screen space. Searches with multiple selections are only supported in
the Advanced Search form and Saved Searches.

In case of a search with multiple selections, include a special option
in the respective search fields to keep the current selection.