Echo end date search fields in quick filter form
The quick filter form defines hidden input fields for the previous
values of date search options, but not for the end date ("between"
filter type). This causes an exception when saving the search profile.

Add the missing fields for end dates. To reduce code duplication, use a
Smarty section to loop through the various date fields.